Reformatted styles
