+++ src/components/Footer.jsx
@@ -1,7 +1,7 @@
 export default function Footer(){
 
     return(
-        <div className="flex justify-end bg-gray-400/20 h-10 items-center">
+        <div className="flex justify-end bg-gray-400/20 h-10 items-center w-screen">
             <p className="text-sm mr-6">
                 Copy right 2022
             </p>
